refactor(#181): features/todo refactor
