+++ src/features/goal/hooks/useGoalId.ts
@@ -1,13 +1,1 @@
-"use client";
-
-import { useParams } from "next/navigation";
-
-export const useGoalId = () => {
-  const { goalId } = useParams();
-
-  if (!goalId) {
-    throw new Error("Goal ID is required");
-  }
-
-  return String(goalId);
-};
+export { useGoalId } from "@/shared/hooks/useGoalId";
