+++ src/features/todo/hooks/useTodoCreateModal.tsx
@@ -7,8 +7,8 @@ import { userQueries } from "@/entities/auth/query/user.queryKey";
 import { goalQueryOptions } from "@/entities/goal";
 import type { Member } from "@/entities/team";
 import { teamQueryOptions } from "@/entities/team";
-import { useGoalId } from "@/features/goal/hooks/useGoalId";
 import { TodoCreateModal } from "@/features/todo/ui/CreateTodoModal/TodoCreateModal";
+import { useGoalId } from "@/shared/hooks/useGoalId";
 import { useOverlay } from "@/shared/hooks/useOverlay";
 
 const TODO_CREATE_MODAL_ID = "todo-create-modal";
