+++ src/shared/hooks/useGoalId/index.ts
@@ -1,0 +1,13 @@
+"use client";
+
+import { useParams } from "next/navigation";
+
+export const useGoalId = () => {
+  const { goalId } = useParams();
+
+  if (!goalId) {
+    throw new Error("Goal ID is required");
+  }
+
+  return String(goalId);
+};
